[KZN-3415] refactors moving out positioning styles
