+++ packages/components/src/__alpha__/SingleSelect/subcomponents/Popover/Popover.module.css
@@ -1,11 +1,6 @@
 @layer kz-components {
   .popover {
     position: absolute;
-    position-anchor: var(--position-anchor, --trigger);
-    top: var(--popover-top, calc(anchor(bottom) + 4px));
-    bottom: var(--popover-bottom, auto);
-    inset-inline-start: var(--popover-inline-start, anchor(start));
-    max-height: var(--popover-max-height, 300px);
     height: auto;
     background-color: var(--color-white);
     border-radius: var(--spacing-8);
@@ -14,11 +9,14 @@
     overflow: hidden auto;
     margin: 0;
     box-sizing: border-box;
-
-    /* TODO: update width based on design */
-    width: 200px;
+    width: 200px; /* TODO: update width based on design */
     visibility: hidden;
     @supports (anchor-name: --anchor) {
+      position-anchor: var(--position-anchor, --trigger);
+      top: var(--popover-top, calc(anchor(bottom) + 4px));
+      bottom: var(--popover-bottom, auto);
+      inset-inline-start: var(--popover-inline-start, anchor(start));
+      max-height: var(--popover-max-height, 300px);
       position-area: bottom center;
       position-try-fallbacks: flip-block;
       width: anchor-size(width);
