+++ packages/components/src/__alpha__/SingleSelect/subcomponents/Popover/Popover.tsx
@@ -1,111 +1,14 @@
-import React, { useEffect, useLayoutEffect, useMemo, useState, type PropsWithChildren } from 'react'
-import { useLocale } from '@react-aria/i18n'
+import React, { useLayoutEffect, useMemo, type PropsWithChildren } from 'react'
+
 import { Popover as RACPopover } from 'react-aria-components'
 import { useSingleSelectContext } from '../../context'
-import { supportsAnchorPositioning, useFixedOverlayPosition } from '../../utils'
+import { usePositioningStyles } from './utils/usePositioningStyles'
 import styles from './Popover.module.css'
 
 type PopoverProps = {
   buttonRef: React.RefObject<HTMLElement>
   popoverRef: React.RefObject<HTMLDivElement>
   racPopoverRef: React.Ref<any>
-}
-
-// CSS custom property names for consistent usage
-const CSS_PROPS = {
-  POSITION_ANCHOR: '--position-anchor',
-  POPOVER_TOP: '--popover-top',
-  POPOVER_BOTTOM: '--popover-bottom',
-  POPOVER_INLINE_START: '--popover-inline-start',
-  POPOVER_MAX_HEIGHT: '--popover-max-height',
-} as const
-
-// Default values
-const DEFAULTS = {
-  MAX_HEIGHT: '300px',
-  OFFSET: '4px',
-} as const
-
-// Cache anchor positioning support detection (SSR-safe)
-const cachedAnchorSupport: boolean | null = null
-
-/**
- * Generates manual positioning styles for browsers without anchor positioning support or SSR
- */
-const getManualPositioningStyles = (positionData: {
-  top: number | string | undefined
-  bottom: number | string | undefined
-  insetInlineStart: number | string | undefined
-  maxHeight: number | string | undefined
-}): React.CSSProperties => ({
-  top: positionData.top,
-  bottom: positionData.bottom,
-  insetInlineStart: positionData.insetInlineStart,
-  maxHeight: positionData.maxHeight,
-  left: 'auto',
-  right: 'auto',
-  position: 'fixed',
-})
-
-/**
- * Generates anchor positioning styles with CSS custom properties
- */
-const getAnchorPositioningStyles = (
-  anchorName: string,
-  positionData: {
-    top: number | string | undefined
-    bottom: number | string | undefined
-    insetInlineStart: number | string | undefined
-    maxHeight: number | string | undefined
-  },
-): React.CSSProperties => {
-  const { maxHeight } = positionData
-  const styles: React.CSSProperties = {
-    [CSS_PROPS.POSITION_ANCHOR]: anchorName,
-  }
-
-  // Max height
-  if (typeof maxHeight === 'number') {
-    styles[CSS_PROPS.POPOVER_MAX_HEIGHT] = `${maxHeight}px`
-  } else if (maxHeight) {
-    styles[CSS_PROPS.POPOVER_MAX_HEIGHT] = maxHeight
-  } else {
-    styles[CSS_PROPS.POPOVER_MAX_HEIGHT] = DEFAULTS.MAX_HEIGHT
-  }
-
-  return styles
-}
-
-/**
- * Hook to determine anchor positioning support and return appropriate styles (memoized)
- */
-const usePopoverPositioning = (
-  anchorName: string,
-  positionData: {
-    top: number | string | undefined
-    bottom: number | string | undefined
-    insetInlineStart: number | string | undefined
-    maxHeight: number | string | undefined
-  },
-): { getPopoverStyle: React.CSSProperties } => {
-  const [hasAnchorSupport, setHasAnchorSupport] = useState<boolean | null>(null)
-
-  useEffect(() => {
-    setHasAnchorSupport(supportsAnchorPositioning(cachedAnchorSupport))
-  }, [])
-
-  // Memoize styles to prevent unnecessary recalculations
-  const getPopoverStyle = useMemo(() => {
-    // Use manual positioning during detection or when not supported
-    if (hasAnchorSupport === null || !hasAnchorSupport) {
-      return getManualPositioningStyles(positionData)
-    }
-
-    // Use CSS anchor positioning when supported
-    return getAnchorPositioningStyles(anchorName, positionData)
-  }, [hasAnchorSupport, anchorName, positionData])
-
-  return { getPopoverStyle }
 }
 
 export const Popover = ({
@@ -115,30 +18,8 @@ export const Popover = ({
   children,
 }: PopoverProps & PropsWithChildren): React.ReactElement => {
   const { isOpen, setOpen, anchorName } = useSingleSelectContext()
-  const { direction } = useLocale()
 
-  // Calculate position for RTL & iframe compatibility
-  const { top, bottom, insetInlineStart, maxHeight, isPositioned } = useFixedOverlayPosition({
-    triggerRef: buttonRef,
-    popoverRef,
-    direction,
-    offset: 4,
-    preferredPlacement: 'bottom',
-  })
-
-  // Memoize position data to prevent unnecessary style recalculations
-  const positionData = useMemo(
-    () => ({
-      top,
-      bottom,
-      insetInlineStart,
-      maxHeight,
-    }),
-    [top, bottom, insetInlineStart, maxHeight],
-  )
-
-  // Generate appropriate styles based on browser support
-  const { getPopoverStyle } = usePopoverPositioning(anchorName, positionData)
+  const { popoverStyle, isPositioned } = usePositioningStyles(buttonRef, popoverRef, anchorName)
 
   // Show/hide popover based on state (memoized for performance, SSR-safe)
   const shouldShowPopover = useMemo(() => isOpen && isPositioned, [isOpen, isPositioned])
@@ -165,7 +46,7 @@ export const Popover = ({
           popover="manual"
           ref={popoverRef}
           className={styles.popover}
-          style={getPopoverStyle}
+          style={popoverStyle}
         >
           {children}
         </div>
