+++ packages/components/src/__alpha__/SingleSelect/subcomponents/Popover/utils/supportsAnchorPositioning.ts
